Move extenstion to shared file
